+++ src/components/Button.tsx
@@ -1,21 +1,24 @@
+import { ButtonSize, ButtonType } from '@/styles/button';
+import { BoldType } from '@/styles/typography';
 import styled from '@emotion/styled';
 import { buttonTypeMap, buttonSizeMap, buttonActiveMap } from '@/styles/button';
 import { bolderMap } from '@/styles/typography';
-import { ButtonProp } from '.';
 import { LayoutMap } from '@/styles/layout';
 import { css } from '@emotion/react';
+import { ButtonHTMLAttributes } from 'react';
 
-export const Container = styled.div`
-  ${LayoutMap.displayFlex}
-  width: fit-content;
-  gap: 10px;
-`;
+export interface ButtonProp extends ButtonHTMLAttributes<HTMLButtonElement> {
+  buttonType?: ButtonType;
+  size?: ButtonSize;
+  bold?: BoldType;
+  fontSize?: number;
+  $active?: boolean;
+}
 
-export const Button = styled.button<ButtonProp>`
+const Button = styled.button<ButtonProp>`
   ${LayoutMap.displayFlex}
   justify-content: center;
   transition: all 0.3s ease;
-  flex: 1;
   ${({ size = 'md' }) => buttonSizeMap[size]}
   ${({ buttonType = 'reverseRadius' }) => css`
     ${buttonTypeMap[buttonType]};
@@ -34,3 +37,5 @@ export const Button = styled.button<ButtonProp>`
         `
       : css``};
 `;
+
+export default Button;
